// src/API.ts
export type User = {
  __typename: "User",
  id: string,
  username: string,
  name?: string | null,
  rating?: number | null,
  avatar?: string | null,
  email: string,
  orders?: ModelOrderConnection | null,
  car?: Car | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelOrderConnection = {
  __typename: "ModelOrderConnection",
  items:  Array<Order | null >,
  nextToken?: string | null,
};

export type Order = {
  __typename: "Order",
  id: string,
  userId: string,
  user?: User | null,
  carId?: string | null,
  car?: Car | null,
  type: string,
  status: string,
  originLat: number,
  originLong: number,
  destLat: number,
  destLong: number,
  createdAt: string,
  updatedAt: string,
};

export type Car = {
  __typename: "Car",
  id: string,
  type: string,
  uri: string,
  latitude?: number | null,
  longitude?: number | null,
  heading?: number | null,
  isAvailable?: boolean | null,
  orders?: ModelOrderConnection | null,
  user?: User | null,
  createdAt: string,
  updatedAt: string,
};